made parser into server extension and updated analysis package accordingly
