+++ src/jupyter-hooks/file-manager.ts
@@ -1,22 +1,17 @@
 import { PathExt } from "@jupyterlab/coreutils";
 
-import { VerNotebook } from "./components/notebook";
-
-import { nbformat } from "@jupyterlab/coreutils";
+import { VerNotebook } from "../components/notebook";
 
 import { Contents, ContentsManager } from "@jupyterlab/services";
 
 import { IDocumentManager } from "@jupyterlab/docmanager";
 
-import { CellRunData } from "./model/checkpoint";
+import { CellRunData } from "../model/checkpoint";
 
-import { NodeyCode } from "./model/nodey";
-
-import { History } from "./model/history";
+import { History } from "../model/history";
 
 export class FileManager {
   readonly docManager: IDocumentManager;
-  private ghostPath: string;
 
   constructor(docManager: IDocumentManager) {
     this.docManager = docManager;
@@ -53,69 +48,6 @@ export class FileManager {
         .then(() => {
           console.log("Model written to file", saveModel);
           accept();
-        })
-        .catch(rej => {
-          //here when you reject the promise if the filesave fails
-          console.error(rej);
-          reject();
-        });
-    });
-  }
-
-  public async openGhost(
-    data: nbformat.INotebookContent,
-    notebook: VerNotebook
-  ): Promise<boolean> {
-    let wasOpen = true;
-    if (!this.ghostPath) {
-      let path = "";
-      var name = notebook.name;
-      path = notebook.path;
-      name = name.substring(0, name.indexOf(".")) + ".ghost";
-      path = "/" + path.substring(0, path.lastIndexOf("/") + 1) + name;
-      await this.writeGhostFile(notebook, data);
-      this.ghostPath = path;
-      wasOpen = false;
-    }
-
-    //let widget = this.docManager.findWidget(path);
-    let widget = this.docManager.openOrReveal(this.ghostPath);
-    if (widget) {
-      console.log("ATTEMPTING TO OPEN GHOST", widget);
-      //(widget.content as GhostBook).feedNewData(data);
-    }
-
-    return wasOpen;
-  }
-
-  private writeGhostFile(notebook: VerNotebook, data: {}): Promise<string> {
-    return new Promise((accept, reject) => {
-      var notebookPath = notebook.path;
-      //console.log("notebook path is", notebookPath)
-      var name = PathExt.basename(notebookPath);
-      name = name.substring(0, name.indexOf(".")) + ".ghost";
-      //console.log("name is", name)
-      var path =
-        "/" +
-        notebookPath.substring(0, notebookPath.lastIndexOf("/") + 1) +
-        name;
-      //console.log("goal path is ", path)
-
-      var saveModel = new HistorySaveModel(
-        name,
-        path,
-        "today",
-        "today",
-        JSON.stringify(data, null, 2)
-      );
-      //console.log("Model to save is", saveModel)
-
-      let contents = new ContentsManager();
-      contents
-        .save(path, saveModel)
-        .then(() => {
-          console.log("Model written to file", saveModel);
-          accept(path);
         })
         .catch(rej => {
           //here when you reject the promise if the filesave fails
@@ -188,36 +120,7 @@ export interface serialized_NodeyOutput extends serialized_Nodey {
   raw: { [key: string]: any };
 }
 
-export interface serialized_NodeyMarkdown extends serialized_Nodey {
-  markdown: string;
-}
-
-export interface serialized_NodeyCode extends serialized_Nodey {
-  type: string;
-  output?: string[];
-  literal?: any;
-  start?: { line: number; ch: number };
-  end?: { line: number; ch: number };
-  content?: any[];
-  right: string;
-}
-
-export interface serialized_NodeyCodeCell extends serialized_NodeyCode {
-  starNodes?: NodeyCode[];
-}
-
 export type serialized_Run = (string | number | CellRunData | string[])[];
-
-export interface serialized_Star {
-  target_type: string;
-  target: string;
-}
-
-export interface serialized_Note {
-  target_type: string;
-  target: string;
-  note: string;
-}
 
 export interface serialized_NodeyHistory {
   runs: serialized_Run[];
@@ -225,6 +128,4 @@ export interface serialized_NodeyHistory {
   nodey: { nodey: number; versions: serialized_Nodey[] }[];
   output: { output: number; versions: serialized_NodeyOutput[] }[];
   deletedCells: number[];
-  stars: serialized_Star[];
-  notes: serialized_Note[];
 }
